Move the toast invocation into the FiltersUpdatedToast module
